Remove OSS tests because there's no SDK on OSS version
